int max implementation for binary

+++ binary.c
@@ -3,7 +3,7 @@
  * binary_div - converts a number to base 2 by recursive split
  * @n: number
  */
-void binary_div(int n)
+void binary_div(long int n)
 {
 	if (n / 2)
 		binary_div(n / 2);
@@ -13,7 +13,7 @@ void binary_div(int n)
  * negative_binary_div - comverts negative decimal to binary
  * @n: number
  */
-void negative_binary_div(int n)
+void negative_binary_div(long int n)
 {
 	if (n / 2)
 		negative_binary_div(n / 2);
@@ -26,7 +26,7 @@ void negative_binary_div(int n)
  */
 int print_binary(va_list list)
 {
-	int n = va_arg(list, int);
+	long int n = va_arg(list, int);
 	int flag = n >= 0 ? 0 : 1;
 
 	n = n > 0 ? n : -n;
